requirements.txt
transformers>=4.46.0
datasets>=3.1.0
torch>=2.4.0
pandas>=2.0.0
scikit-learn>=1.3.0
PyYAML>=6.0
tqdm
psutil
huggingface_hub>=0.20.0
transformers>=4.40.0
safetensors